Refactor mock data creation to MockUtil
